backend/requirements.txt
Django==4.1.3
Pillow==9.3.0
djangorestframework==3.14.0
psycopg2==2.9.5
django-cors-headers==3.13.0